+++ src/BORDL.cpp
@@ -299,7 +299,7 @@ struct BORDL : Module {
 	int SCALE_HARMONIC_MINOR [7] = {0, 2, 3, 5, 7, 8, 11};
 	int SCALE_INDIAN         [7] = {0, 1, 1, 4, 5, 8, 10};
 	int SCALE_LOCRIAN        [7] = {0, 1, 3, 5, 6, 8, 10};
-	int SCALE_LYDIAN         [7] = {0, 2, 4, 6, 7, 9, 10};
+	int SCALE_LYDIAN         [7] = {0, 2, 4, 6, 7, 9, 11};
 	int SCALE_MAJOR          [7] = {0, 2, 4, 5, 7, 9, 11};
 	int SCALE_MELODIC_MINOR  [9] = {0, 2, 3, 5, 7, 8, 9, 10, 11};
 	int SCALE_MINOR          [7] = {0, 2, 3, 5, 7, 8, 10};